# Move testing setup out of sbt/TestsPlugin