--- a/.travis.yml
+++ b/.travis.yml
@@ -7,9 +7,9 @@ jdk:
 
 jobs:
   include:
-  - { name: testFunctional 2.11, script: sbt -Dmima.testScalaVersion=2.11.12 testFunctional }
-  - { name: testFunctional 2.12, script: sbt -Dmima.testScalaVersion=2.12.11 testFunctional }
-  - { name: testFunctional 2.13, script: sbt -Dmima.testScalaVersion=2.13.2  testFunctional }
+  - { name: testFunctional 2.11, script: sbt "functional-tests/runMain com.typesafe.tools.mima.lib.UnitTests -211" }
+  - { name: testFunctional 2.12, script: sbt "functional-tests/runMain com.typesafe.tools.mima.lib.UnitTests -212" }
+  - { name: testFunctional 2.13, script: sbt "functional-tests/runMain com.typesafe.tools.mima.lib.UnitTests -213" }
   - { name: scripted 1/2, script: sbt "scripted sbt-mima-plugin/*1of2" }
   - { name: scripted 2/2, script: sbt "scripted sbt-mima-plugin/*2of2" }