--- a/functional-tests/src/main/scala/com/typesafe/tools/mima/lib/Coursier.scala
+++ b/functional-tests/src/main/scala/com/typesafe/tools/mima/lib/Coursier.scala
@@ -1,0 +1,10 @@
+package com.typesafe.tools.mima.lib
+
+import coursier._
+
+object Coursier {
+  val scalaEA = mvn"https://scala-ci.typesafe.com/artifactory/scala-integration/"
+  val scalaPR = mvn"https://scala-ci.typesafe.com/artifactory/scala-pr-validation-snapshots/"
+
+  def fetch(dep: Dependency) = Fetch().addRepositories(scalaEA, scalaPR).addDependencies(dep).run()
+}