--- a/functional-tests/src/test/class-becomes-trait-nok/app/App.scala
+++ b/functional-tests/src/test/class-becomes-trait-nok/app/App.scala
@@ -1,6 +1,5 @@
 object App {
   def main(args: Array[String]): Unit = {
-    println(new A)
     println(new A().foo)
   }
 }