--- a/functional-tests/src/test/class-method-implementation-result-type-changed-nok/app/App.scala
+++ b/functional-tests/src/test/class-method-implementation-result-type-changed-nok/app/App.scala
@@ -1,3 +1,5 @@
-object Main extends App {
-  println(new Foo().bar)
+object App {
+  def main(args: Array[String]): Unit = {
+    println(new Foo().bar)
+  }
 }