I've been interested in experimenting in the testing space, separating
the critical details of a test from the arbitrary, for instance changing
a `val` to a `def`, but the name doesn't matter.  So, basically,
property-based testing, but for code.

To get to that I first had rework the testing infrastructure in MiMa,
which, as is, was defined in the build-space, using sbt/zinc
infrastructure.  Wanting to move to generated sources/templates, I
reworked it thusly.

For now, I've left the test sources where they are.  Build load time are
much faster (from 20s to 7s) because each test isn't a project any more
(and IDE import will be faster as a consequence).  Test running will be
slower (from 50s to 90s), because I didn't add any concurrency in
running them and they're batch compiled, not incrementally compiled
(cached).  Also unhappy paths will be different (like test source
compilation errors and app run linkage errors), but I'll fix those up as
they show up.

I also changed the integration tests to use Coursier.  Interestingly the
coursier artifact depends on scala-reflect (transitively, through
argonaut).

Additional changes:
* class-becomes-trait-nok: a change from starting to generalise it;
* class-method-implementation-result-type-changed-nok: stick to 'App';
  avoid having to scan for main classes;
* SbtMima unused import;
* I tried defining `testFunctional` as an input with
  `runMain.toPartial`, but it looks like that's broken.